Revert "Add fast BLANC computation"
