+++ setup.cfg
@@ -10,7 +10,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 classifiers =
     License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.6
     Environment :: Console
 keywords =
     nlp
